+++ lorien/UI/MainMenu.gd
@@ -5,30 +5,28 @@ class_name MainMenu
 signal open_about_dialog
 signal open_settings_dialog
 signal open_url(url)
-signal export_as(type)
 signal open_project(filepath)
 signal save_project
 signal save_project_as
+signal export_svg
 
 # -------------------------------------------------------------------------------------------------
 const ITEM_OPEN 		:= 0
 const ITEM_SAVE 		:= 1
 const ITEM_SAVE_AS 		:= 2
-const ITEM_SETTINGS 	:= 3
-const ITEM_MANUAL 		:= 4
-const ITEM_BUG_TRACKER 	:= 5
-const ITEM_ABOUT 		:= 6
+const ITEM_EXPORT 		:= 3
+const ITEM_SETTINGS 	:= 4
+const ITEM_MANUAL 		:= 5
+const ITEM_BUG_TRACKER 	:= 6
+const ITEM_ABOUT 		:= 7
 
 const ITEM_VIEW_1 		:= 100
 const ITEM_VIEW_2 		:= 101
 const ITEM_VIEW_3 		:= 102
 
-const ITEM_EXPORT_PNG	:= 200
-
 # -------------------------------------------------------------------------------------------------
 export var file_dialog_path: NodePath
 onready var _submenu_views: PopupMenu = $ViewsMenu
-onready var _submenu_export: PopupMenu = $ExportMenu
 
 # -------------------------------------------------------------------------------------------------
 func _ready():
@@ -37,16 +35,11 @@ func _ready():
 	_submenu_views.add_item("View 1", ITEM_VIEW_1)
 	_submenu_views.add_item("View 2", ITEM_VIEW_2)
 	
-	# Export submenu
-	_submenu_export.name = tr("MENU_EXPORT")
-	_submenu_export.add_item(tr("MENU_EXPORT_PNG"), ITEM_EXPORT_PNG)
-	
 	# main menu
 	add_item(tr("MENU_OPEN"), ITEM_OPEN)
 	add_item(tr("MENU_SAVE"), ITEM_SAVE)
 	add_item(tr("MENU_SAVE_AS"), ITEM_SAVE_AS)
-	add_submenu_item(_submenu_export.name, tr("MENU_EXPORT"))
-	add_separator()
+	add_item(tr("MENU_EXPORT"), ITEM_EXPORT)
 	add_item(tr("MENU_SETTINGS"), ITEM_SETTINGS)
 	add_separator()
 	add_item(tr("MENU_MANUAL"), ITEM_MANUAL)
@@ -59,6 +52,7 @@ func _on_MainMenu_id_pressed(id: int):
 		ITEM_OPEN: _on_open_project()
 		ITEM_SAVE: emit_signal("save_project")
 		ITEM_SAVE_AS: emit_signal("save_project_as")
+		ITEM_EXPORT: emit_signal("export_svg")
 		ITEM_SETTINGS: emit_signal("open_settings_dialog")
 		ITEM_MANUAL: emit_signal("open_url", "https://github.com/mbrlabs/lorien/blob/main/docs/manuals/manual_v0.5.0.md")
 		ITEM_BUG_TRACKER: emit_signal("open_url", "https://github.com/mbrlabs/lorien/issues")
@@ -82,11 +76,6 @@ func _on_file_dialog_closed() -> void:
 	var file_dialog: FileDialog = get_node(file_dialog_path)
 	Utils.remove_signal_connections(file_dialog, "file_selected")
 	Utils.remove_signal_connections(file_dialog, "popup_hide")
-
-# -------------------------------------------------------------------------------------------------
-func _on_SaveAsMenu_id_pressed(id: int):
-	match id:
-		ITEM_EXPORT_PNG: emit_signal("export_as", Types.ExportType.PNG)
 
 # -------------------------------------------------------------------------------------------------
 func add_item_with_shortcut(target: PopupMenu, name: String, id: int, shortcut_action: String) -> void: